fix 32bit bit-shift

+++ rtl837x_regs.h
@@ -308,7 +308,7 @@
 #ifdef REGDBG
 
 #define REG_SET(r, v) do { \
-	SFR_DATA_U16_UPPER = (uint16_t)(v >> 16); \
+	SFR_DATA_U16_UPPER = (uint16_t)((uint32_t)(v) >> 16); \
 	SFR_DATA_U16 = (uint16_t)(v); \
 	reg_write(r); \
 	write_char('R'); print_byte(r >> 8); print_byte(r); write_char('-'); \
@@ -324,8 +324,9 @@
 	write_char('R'); print_byte(r>>8); print_byte(r); write_char('-'); print_byte(v24); print_byte(v16); print_byte(v8); print_byte(v0); write_char(' '); \
 } while (0)
 #else
+
 #define REG_SET(r, v) do { \
-	SFR_DATA_U16_UPPER = (uint16_t)(v >> 16); \
+	SFR_DATA_U16_UPPER = (uint16_t)((uint32_t)(v) >> 16); \
 	SFR_DATA_U16 = (uint16_t)(v); \
 	reg_write(r); \
 } while (0)
